feat: split out environments into own files
