+++ libs/@local/hashql/core/src/type/environment/context/auxiliary.rs
