+++ libs/@local/hashql/core/src/type/mod.rs
@@ -3,6 +3,7 @@
 mod collection;
 pub mod environment;
 pub mod error;
+pub mod infer;
 pub mod intern;
 pub mod kind;
 pub mod lattice;
